+++ src/Provider/Klarna/RequestParameter/Enricher/OrderLinesRequestParameterEnricher.php
@@ -6,10 +6,19 @@ namespace PayonePayment\Provider\Klarna\RequestParameter\Enricher;
 
 use PayonePayment\Components\Hydrator\LineItemHydrator\LineItemHydratorInterface;
 use PayonePayment\RequestParameter\Enricher\OrderLinesRequestParameterEnricherTrait;
+use PayonePayment\RequestParameter\PaymentRequestDto;
 use PayonePayment\RequestParameter\RequestParameterEnricherInterface;
 
+/**
+ * @template T as PaymentRequestDto
+ *
+ * @implements RequestParameterEnricherInterface<T>
+ */
 readonly class OrderLinesRequestParameterEnricher implements RequestParameterEnricherInterface
 {
+    /**
+     * @use OrderLinesRequestParameterEnricherTrait<T>
+     */
     use OrderLinesRequestParameterEnricherTrait;
 
     public function __construct(